+++ src/Equation.hpp
@@ -14,11 +14,11 @@ using namespace std;
 
 struct Equation
 {
-    const int N;
+    const size_t N;
     SparseMatrix<float> K;
     vector<float> f;
 
-    Equation(const int N);
+    Equation(const size_t N);
     pair<unique_ptr<vector<float>>, int> SolveIterative();
     void Print();
     friend ostream& operator<<(ostream& os, const Equation& equation);
